requirements.txt
inquirer
PyYAML
watchdog
